# Fix String interpolation optimisation for VWA